+++ src/App.tsx
@@ -1,5 +1,4 @@
-import { Home, Plan, Discover, Check, NoPage } from "./Pages";
-import { Navbar } from "./componets";
+import { Home, Plan, Discover, Check, NoPage, Kitchen } from "./Pages";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import "./App.css";
 
@@ -7,17 +6,15 @@ function App() {
   return (
     <div className="App">
       <BrowserRouter>
-        <Navbar />
-        <div className="pages">
-          <Routes>
-            <Route index element={<Home />} />
-            <Route path="/home" element={<Home />} />
-            <Route path="/planatrip" element={<Plan />} />
-            <Route path="/discover" element={<Discover />} />
-            <Route path="/checkflights" element={<Check />} />
-            <Route path="*" element={<NoPage />} />
-          </Routes>
-        </div>
+        <Routes>
+          <Route index element={<Home />} />
+          <Route path="/home" element={<Home />} />
+          <Route path="/planatrip" element={<Plan />} />
+          <Route path="/discover" element={<Discover />} />
+          <Route path="/checkflights" element={<Check />} />
+          <Route path="*" element={<NoPage />} />
+          <Route path="/kitchen" element={<Kitchen />} />
+        </Routes>
       </BrowserRouter>
     </div>
   );
