+++ src/Pages/home/home.tsx
@@ -1,10 +1,13 @@
-import React from "react";
+import { Navbar } from "../../components";
 
 const Home = () => {
   return (
-    <div className="text-3xl w-screen h-screen flex items-center justify-center bg-yellow-300">
-      Home Section
-    </div>
+    <>
+      <Navbar />
+      <div className="text-3xl w-screen h-screen flex items-center justify-center bg-yellow-300">
+        Home Section
+      </div>
+    </>
   );
 };
