+++ src/Pages/index.tsx
@@ -4,3 +4,6 @@ export { default as Plan } from "./plan/plan";
 export { default as Discover } from "./discover/discover";
 export { default as Check } from "./check/check";
 export { default as NoPage } from "./nopage/noPage";
+
+// Kitchen sink to export to app.tsx
+export { default as Kitchen } from "./kitchen_sink/kitchen_sink";
