+++ src/Pages/plan/plan.tsx
@@ -1,10 +1,13 @@
-import React from "react";
+import { Navbar } from "../../components";
 
 const plan = () => {
   return (
-    <div className="text-3xl w-screen h-screen flex items-center justify-center bg-cyan-500">
-      Plan a trip page
-    </div>
+    <>
+      <Navbar />
+      <div className="text-3xl w-screen h-screen flex items-center justify-center bg-cyan-500">
+        Plan a trip page
+      </div>
+    </>
   );
 };
