+++ tailwind.config.js
@@ -2,7 +2,12 @@
 module.exports = {
   content: ["./src/**/*.{js,jsx,ts,tsx}"],
   theme: {
-    extend: {},
+    extend: {
+      fontFamily: {
+        noto: ["Noto Sans", "sans - serif"],
+        robo: ["Roboto", "sans - serif"],
+      },
+    },
   },
   plugins: [],
 };
